Moved authentication features to a separate app
