+++ src/hr_management/authentication/templates/login.html
@@ -1,0 +1,59 @@
+{% load static %}
+
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" type="text/css" href="{% static 'base.css' %}">
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-GLhlTQ8iRABdZLl6O3oVMWSktQOp6b7In1Zl3/Jr59b6EGGoI1aFkw7cmDA6j6gD" crossorigin="anonymous">
+    <link rel="shortcut icon" type="image/png" href="{% static 'images/favicon.png' %}">
+    <title>Home</title> 
+</head>
+
+
+    <!-- Print messages -->
+    {% if messages %}
+        {% for message in messages %}
+            <div class="alert alert-warning alert-dismissible fade show" role="alert">
+                {{ message }}
+            <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>
+            </div>
+        {% endfor %}
+    {% endif %}
+
+    <div>
+        <div>
+            <br/><br/>
+            <div class="logo-container">
+                <img src="{% static 'images/penguin_logo.png' %}" alt="Application Logo" class="navbar-logo">
+            </div>
+                <br/><br/>
+                <div class="col-md-6 offset-md-3">
+                <h1>{{title}}</h1>
+                <br/>
+                <form method="POST" action="{% url 'authentication:login' %}">
+                    {% csrf_token %} <!-- ("Cross-Site Request Forgery Token") -->
+            
+                    <div class="mb-3">
+                    <input type="text" class="form-control" name="email", placeholder="email" required>
+                    
+                    </div><br/>
+                    <div class="mb-3">
+                    <input type="password" class="form-control" name="password", placeholder="Password" required>
+                    </div>
+                    <br/>
+                    <button type="submit" class="btn btn-secondary">Login</button>
+                </form>
+                </div>
+        </div>
+            
+        <!-- Button on the bottom of page opens user manual-->
+        <button class="help-button">?</button>
+    </div>
+
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/js/bootstrap.bundle.min.js" integrity="sha384-w76AqPfDkMBDXo30jS1Sgez6pr3x5MlQ1ZAGC+nuZB+EYdgRZgiwxhTBTkF7CXvN" crossorigin="anonymous"></script>
+
+</body>
+</html>
+
