+++ src/hr_management/authentication/urls.py
@@ -1,0 +1,10 @@
+from django.urls import path
+from . import views
+
+
+app_name = 'authentication'
+
+urlpatterns = [
+    path('', views.login_user, name="login"),
+    path('logout', views.logout_user, name='logout'),
+]
